data-plane/python/integrations/slimrpc/pyproject.toml
# Copyright AGNTCY Contributors (https://github.com/agntcy)
# SPDX-License-Identifier: Apache-2.0

[project]
name = "slimrpc"
version = "0.1.0"
description = "RPC over SLIM Library"
readme = "README_pypi.md"
requires-python = ">=3.9"
license = "Apache-2.0"
dependencies = [
    "slim-bindings>=0.5.0",
    "googleapis-common-protos>=1.70.0",
    "grpcio>=1.74.0",
    "async-timeout>=5.0.1",
]

[dependency-groups]
linting = ["ruff>=0.12", "mypy>=1.17.0", "types-protobuf>=6.30.2.20250822"]
testing = ["pytest>=8.3.4", "pytest-asyncio>=0.24.0", "pytest-cov>=6.0.0"]

[tool.uv]
default-groups = ["linting", "testing"]

# [tool.uv.sources]
# slim-bindings = { path = "../../bindings" }

[project.scripts]
simple-server = "slimrpc.examples.simple.server:main"
simple-client = "slimrpc.examples.simple.client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["slimrpc"]

[tool.ruff.format]
exclude = ["*_pb2.py", "*_pb2.pyi", "*_pb2_slimrpc.py"]

[tool.ruff.lint]
extend-select = ["I", "ASYNC", "SIM", "B", "ANN"]
exclude = ["*_pb2.py", "*_pb2.pyi", "*_pb2_slimrpc.py"]

[[tool.mypy.overrides]]
module = ["google.rpc.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["slim_bindings.*"]
ignore_missing_imports = true
